Very Rough Draft.
I've spent most of this time creating a firebase connection to handle user authentication and shortly, database storage. The design and pages are very temporary, and will change drastically as this moves on.

Tomorrow I will focus on finalizing the database and admin side to allow us to start building the backend, and go from there.