Merge database logic with frontend logic
I took the current homepage and made it an index, and then added the logic to make it function with the backend correctly.

Currently, it redirects (us) to the admin page, but once the user page is done, it will transport a new user to that page to set up their profile, and then add the tags they like to their user.