Created User `profile` page, and refined hobby search page.
I created a user profile page which is very basic right now and very little styling, but wanted to get the functionality there and I think it is. The user is allow to select tags and save to their profile, which then takes them to the Hobby Search, which now has some basic functionality as well. It shows the Hobby name and photo, and if the user clicks on it, it will expand with more info and the user is allowed to add it to their profile.

Need to add logic to allow the user to go back to their profile.